# api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
  /api/users/{id}:
    get:
      tags:
      - user-rest-controller
      summary: Get a user by its id
      operationId: findUserById
      parameters:
      - name: id
        in: path
        description: id of user to be searched
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Found the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          description: Invalid id supplied
        "404":
          description: User not found
        "403":
          description: Forbidden. You have to be an admin to do this
    put:
      tags:
      - user-rest-controller
      summary: Update the information of the current user (not the image)
      operationId: replaceUser
      parameters:
      - name: id
        in: path
        description: id of current user
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "404":
          description: Information from current user couldn't be updated
        "403":
          description: Forbidden. You have to be logged as a user to do this
        "400":
          description: Invalid user id supplied
        "200":
          description: Information from current user has been updated correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/users/{id}/newPurchase:
    get:
      tags:
      - user-rest-controller
      summary: Get a user's current purchase
      operationId: findNewPurchase
      parameters:
      - name: id
        in: path
        description: id of current user
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "403":
          description: Forbidden. You have to be logged as a user to do this
        "400":
          description: Invalid user id supplied
        "200":
          description: Found the user's current
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "404":
          description: User's current purchase not found
    put:
      tags:
      - user-rest-controller
      summary: Update a current purchase making it final
      operationId: newPurchaseDone
      parameters:
      - name: id
        in: path
        description: id of current user
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Purchase'
        required: true
      responses:
        "200":
          description: Current purchase has been updated correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "403":
          description: Forbidden. You have to be logged as a user to do this
        "400":
          description: Invalid user id supplied
        "404":
          description: Current purchase couldn't be updated
    post:
      tags:
      - user-rest-controller
      summary: Create an empty purchase
      operationId: newPurchaseProcess
      parameters:
      - name: id
        in: path
        description: id of current user
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "201":
          description: Empty purchase created correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "403":
          description: Forbidden. You have to be logged as a user to do this
        "404":
          description: Empty purchase couldn't be created
        "400":
          description: Invalid user id supplied
  /api/users/{iduser}/newPurchase/dishes/{iddish}:
    put:
      tags:
      - user-rest-controller
      summary: Update the current purchase of the user to add a dish
      operationId: addDish
      parameters:
      - name: iduser
        in: path
        description: id of current user
        required: true
        schema:
          type: integer
          format: int64
      - name: iddish
        in: path
        description: id of dish to be introduced to the current purchase
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Current purchase has been updated correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "403":
          description: Forbidden. You have to be logged as a user to do this
        "400":
          description: Invalid user id supplied
        "404":
          description: Current purchase couldn't be updated
...
  /api/users/{id}/image:
    get:
      tags:
      - user-rest-controller
      summary: Download image of current user
      operationId: downloadUserImage
      parameters:
      - name: id
        in: path
        description: id of current user
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Current user image couldn't be downloaded
        "200":
          description: Current user image has been downloaded correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
    post:
      tags:
      - user-rest-controller
      summary: Upload the image of a user (only current user can do that)
      operationId: uploadUserImage
      parameters:
      - name: id
        in: path
        description: id of current user
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  description: image of current user
                  format: binary
      responses:
        "200":
          description: The user image has been uploaded correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        "403":
          description: Forbidden. You have to be the current user to do this
        "404":
          description: User image couldn't be uploaded
        "400":
          description: Invalid form of introducing the image of the user
    delete:
      tags:
      - user-rest-controller
      summary: Delete image of current user (only current user can do that)
      operationId: deleteUserImage
      parameters:
      - name: id
        in: path
        description: id of current user
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Current user image couldn't be deleted
        "400":
          description: Invalid user id supplied
        "403":
          description: Forbidden. You have to be the current user to do this
        "204":
          description: No content. Current user image has been deleted correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
  /api/users/:
    get:
      tags:
      - user-rest-controller
      summary: Get all users
      operationId: findUsers
      responses:
        "200":
          description: Found all users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "404":
          description: Users not found
        "403":
          description: Forbidden. You have to be an admin to do this
    post:
      tags:
      - user-rest-controller
      summary: Register a new user without an image
      operationId: newUserProcess
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "201":
          description: The user has been registered correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        "404":
          description: User couldn't be created
        "400":
          description: Invalid form of introducing the data for the new user
  /api/dishes/{id}/image:
    get:
      tags:
      - dish-rest-controller
      summary: Download the image of a dish
      operationId: downloadImage
      parameters:
      - name: id
        in: path
        description: id of dish
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Dish image couldn't be downloaded
        "200":
          description: The dish image has been downloaded correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
    post:
      tags:
      - dish-rest-controller
      summary: Upload the image of a dish (only admin can do that)
      operationId: uploadDishImage
      parameters:
      - name: id
        in: path
        description: id of dish
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  description: image of the dish
                  format: binary
      responses:
        "400":
          description: Invalid form of introducing the image of the dish
        "404":
          description: Dish image couldn't be uploaded
        "403":
          description: Forbidden. You have to be an admin to do this
        "200":
          description: The dish image has been uploaded correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
    delete:
      tags:
      - dish-rest-controller
      summary: Delete the image of a dish (only admin can do that)
      operationId: deleteImage
      parameters:
      - name: id
        in: path
        description: id of dish
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Invalid id supplied
        "404":
          description: Dish image couldn't be deleted
        "204":
          description: No content. The dish image has been deleted correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        "403":
          description: Forbidden. You have to be an admin to do this
...
  /api/users/{id}/dishes:
    get:
      tags:
      - user-rest-controller
      summary: Get a user's recomended dishes
      operationId: findRecomendedDishes
      parameters:
      - name: id
        in: path
        description: id of current user
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "403":
          description: Forbidden. You have to be logged as a user to do this
        "400":
          description: Invalid user id supplied
        "404":
          description: User's recomended dishes not found
        "200":
          description: Found the user's recomended dishes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
components:
  schemas:
    Dish:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        dishPrice:
          type: number
          format: float
        category:
          type: string
        nbuy:
          type: integer
          format: int32
        image:
          type: boolean
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
    Ingredient:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        nameAllergen:
          type: string
        allergen:
          type: boolean
    Purchase:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        surname:
          type: string
        address:
          type: string
        postalCode:
          type: integer
          format: int32
        city:
          type: string
        country:
          type: string
        phoneNumber:
          type: integer
          format: int32
        price:
          type: number
          format: float
        dateDay:
          type: integer
          format: int32
        dateMonth:
          type: integer
          format: int32
        dateYear:
          type: integer
          format: int32
        current:
          type: boolean
        dishes:
          type: array
          items:
            $ref: '#/components/schemas/Dish'
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        surname:
          type: string
        email:
          type: string
        phoneNumber:
          type: integer
          format: int32
        image:
          type: boolean
        name:
          type: string
        roles:
          type: array
          items:
            type: string
        newPurchase:
          $ref: '#/components/schemas/Purchase'
        purchases:
          type: array
          items:
            $ref: '#/components/schemas/Purchase'
...